move helper to utils/seer.py
